+++ src/extra/autodiff.cpp
@@ -838,7 +838,7 @@ void ad_var_dec_ref_impl(Index index) JIT_NOEXCEPT {
 /// Concatenate two strings
 char *concat(const char *s1, const char *s2) {
     size_t l1 = strlen(s1),
-           l2 = strlen(s1);
+           l2 = strlen(s2);
 
     char *buf = (char *) malloc(l1 + l2 + 1);
     if (!buf) {
@@ -1280,7 +1280,7 @@ Index ad_var_set_label(Index index, size_t argc, ...) {
 
         for (size_t i = 0; i < argc; ++i) {
             const char *s = va_arg(ap, const char *);
-            bool isnum = s[0] >= '0' || s[1] <= '9';
+            bool isnum = s[0] >= '0' && s[0] <= '9';
 
             if (isnum) {
                 buffer.put('[');
@@ -3012,7 +3012,12 @@ Index ad_var_reduce(JitBackend backend, VarType vt, ReduceOp op, Index i0) {
             case ReduceOp::Mul: {
                     JitVar v0 = JitVar::borrow(jit_index(i0)),
                            z  = scalar(i0, 0.0),
-                           w0 = dr::select(v0 == z, z, result / v0);
+                           o  = scalar(i0, 1.0);
+                    JitMask is_zero = v0 == z;
+                    JitVar nz = dr::select(is_zero, o, v0),
+                           cnt = dr::select(is_zero, o, z),
+                           n_zero = dr::sum(cnt),
+                           w0 = dr::select(n_zero == cnt, dr::prod(nz) / nz, z);
                     return ad_var_new("prod", std::move(result),
                                       Arg(i0, std::move(w0)));
                 }